tests: refactor tests utils
