+++ tests/conftest.py
@@ -2,9 +2,7 @@ import os
 from uuid import uuid4
 
 import pytest
-from fed_reg.provider.enum import ProviderType
 from fed_reg.provider.schemas_extended import (
-    AuthMethodCreate,
     BlockStorageServiceCreateExtended,
     ComputeServiceCreateExtended,
     IdentityProviderCreateExtended,
@@ -14,20 +12,24 @@ from fed_reg.provider.schemas_extended import (
     ProjectCreate,
     ProviderCreateExtended,
     RegionCreateExtended,
-    SLACreateExtended,
-    UserGroupCreateExtended,
 )
 from fed_reg.service.enum import ObjectStoreServiceName
-from pytest_cases import parametrize
 
-from tests.schemas.utils import (
+from tests.fed_reg.utils import (
+    fed_reg_provider_dict,
     random_block_storage_service_name,
     random_compute_service_name,
     random_identity_service_name,
-    random_lower_string,
     random_network_service_name,
-    random_url,
+)
+from tests.schemas.utils import (
+    auth_method_dict,
     sla_dict,
+    user_group_dict,
+)
+from tests.utils import (
+    random_lower_string,
+    random_url,
 )
 
 
@@ -35,6 +37,19 @@ from tests.schemas.utils import (
 def clear_os_environment() -> None:
     """Clear the OS environment."""
     os.environ.clear()
+
+
+@pytest.fixture
+def identity_provider_create() -> IdentityProviderCreateExtended:
+    """Fixture with an IdentityProviderCreateExtended."""
+    auth_method = auth_method_dict()
+    auth_method["idp_name"] = auth_method.pop("name")
+    return IdentityProviderCreateExtended(
+        user_groups=[{**user_group_dict(), "sla": {**sla_dict(), "project": uuid4()}}],
+        endpoint=random_url(),
+        group_claim=random_lower_string(),
+        relationship=auth_method,
+    )
 
 
 @pytest.fixture
@@ -92,66 +107,6 @@ def s3_service_create() -> ObjectStoreServiceCreateExtended:
 
 
 @pytest.fixture
-@parametrize(
-    s=[
-        block_storage_service_create,
-        compute_service_create,
-        identity_service_create,
-        network_service_create,
-    ]
-)
-def service_create(
-    s: BlockStorageServiceCreateExtended
-    | ComputeServiceCreateExtended
-    | IdentityServiceCreate
-    | NetworkServiceCreateExtended,
-) -> (
-    BlockStorageServiceCreateExtended
-    | ComputeServiceCreateExtended
-    | IdentityServiceCreate
-    | NetworkServiceCreateExtended
-):
-    """Parametrized fixture with all possible services.
-
-    BlockStorageServiceCreateExtended, ComputeServiceCreateExtended,
-    IdentityServiceCreate and NetworkServiceCreateExtended.
-    """
-    return s
-
-
-@pytest.fixture
-def sla_create() -> SLACreateExtended:
-    """Fixture with an SLACreateExtended."""
-    return SLACreateExtended(**sla_dict(), project=uuid4())
-
-
-@pytest.fixture
-def user_group_create(sla_create: SLACreateExtended) -> UserGroupCreateExtended:
-    """Fixture with a UserGroupCreateExtended."""
-    return UserGroupCreateExtended(name=random_lower_string(), sla=sla_create)
-
-
-@pytest.fixture
-def auth_method_create() -> AuthMethodCreate:
-    return AuthMethodCreate(
-        idp_name=random_lower_string(), protocol=random_lower_string()
-    )
-
-
-@pytest.fixture
-def identity_provider_create(
-    auth_method_create: AuthMethodCreate, user_group_create: UserGroupCreateExtended
-) -> IdentityProviderCreateExtended:
-    """Fixture with an IdentityProviderCreateExtended."""
-    return IdentityProviderCreateExtended(
-        user_groups=[user_group_create],
-        endpoint=random_url(),
-        group_claim=random_lower_string(),
-        relationship=auth_method_create,
-    )
-
-
-@pytest.fixture
 def region_create(
     block_storage_service_create: BlockStorageServiceCreateExtended,
     compute_service_create: ComputeServiceCreateExtended,
@@ -178,8 +133,7 @@ def provider_create(
 ):
     identity_provider_create.user_groups[0].sla.project = project_create.uuid
     return ProviderCreateExtended(
-        name=random_lower_string(),
-        type=ProviderType.OS,
+        **fed_reg_provider_dict(),
         identity_providers=[identity_provider_create],
         regions=[region_create],
         projects=[project_create],
