+++ tests/providers/common/test_conn_thred.py
@@ -12,11 +12,11 @@ from tests.schemas.utils import (
     issuer_dict,
     openstack_dict,
     project_dict,
-    random_lower_string,
     region_dict,
     sla_dict,
     user_group_dict,
 )
+from tests.utils import random_lower_string
 
 
 class CaseProvider:
