+++ tests/providers/common/test_get_provider.py
@@ -20,10 +20,10 @@ from tests.schemas.utils import (
     kubernetes_dict,
     openstack_dict,
     project_dict,
-    random_lower_string,
     sla_dict,
     user_group_dict,
 )
+from tests.utils import random_lower_string
 
 
 class CaseProviderStatus:
