+++ tests/providers/common/test_get_provider_components.py
@@ -22,10 +22,10 @@ from tests.schemas.utils import (
     kubernetes_dict,
     openstack_dict,
     project_dict,
-    random_lower_string,
     sla_dict,
     user_group_dict,
 )
+from tests.utils import random_lower_string
 
 
 @patch("src.providers.conn_thread.OpenstackData")
