+++ tests/providers/common/test_get_updated_idp.py
@@ -11,10 +11,10 @@ from tests.schemas.utils import (
     issuer_dict,
     openstack_dict,
     project_dict,
-    random_lower_string,
     sla_dict,
     user_group_dict,
 )
+from tests.utils import random_lower_string
 
 
 class CaseProviderThread:
