+++ tests/providers/common/test_prepare_project_conf.py
@@ -9,10 +9,10 @@ from tests.schemas.utils import (
     openstack_dict,
     private_net_proxy_dict,
     project_dict,
-    random_lower_string,
     sla_dict,
     user_group_dict,
 )
+from tests.utils import random_lower_string
 
 
 class CaseProviderThread:
