+++ tests/providers/openstack/conftest.py
@@ -9,8 +9,8 @@ from tests.schemas.utils import (
     auth_method_dict,
     openstack_dict,
     project_dict,
-    random_lower_string,
 )
+from tests.utils import random_lower_string
 
 
 @pytest.fixture
