+++ tests/providers/openstack/test_block_storage_services.py
@@ -12,7 +12,7 @@ from pytest_cases import parametrize_with_cases
 
 from src.models.provider import Limits
 from src.providers.openstack import OpenstackData
-from tests.schemas.utils import random_url
+from tests.utils import random_url
 
 
 class CaseEndpointResp:
