+++ tests/providers/openstack/test_object_store_services.py
@@ -13,7 +13,7 @@ from pytest_cases import parametrize_with_cases
 
 from src.models.provider import Limits
 from src.providers.openstack import OpenstackData
-from tests.schemas.utils import random_url
+from tests.utils import random_url
 
 
 class CaseEndpointResp:
