+++ tests/providers/openstack/test_s3_services.py
@@ -11,7 +11,7 @@ from pytest_cases import parametrize_with_cases
 
 from src.models.provider import Limits
 from src.providers.openstack import OpenstackData
-from tests.schemas.utils import random_lower_string, random_url
+from tests.utils import random_lower_string, random_url
 
 
 class CaseEndpointResp:
