+++ tests/schemas/test_openstack.py
@@ -9,9 +9,9 @@ from tests.schemas.utils import (
     auth_method_dict,
     openstack_dict,
     project_dict,
-    random_lower_string,
     region_dict,
 )
+from tests.utils import random_lower_string
 
 # TODO: Add BlockStorageVolMap case
