+++ tests/test_utils.py
@@ -22,10 +22,10 @@ from tests.schemas.utils import (
     kubernetes_dict,
     openstack_dict,
     project_dict,
-    random_lower_string,
     sla_dict,
     user_group_dict,
 )
+from tests.utils import random_lower_string
 
 
 class CaseYamlFiles:
